% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bulk_sample}
\alias{bulk_sample}
\title{Create pseudo-bulk for each sample in a Seurat object}
\usage{
bulk_sample(data, ncores = 1)
}
\arguments{
\item{data}{A Seurat object containing expression.}

\item{ncores}{Number of cores for parallel computation.}
}
\value{
A Seurat object containing the pseudo-bulks for each sample
in the input data.
}
\description{
This function takes a Seurat object as input and creates
a pseudo-bulk for each sample in the object. This is calculated by
taking row means of the expression for each sample.
}
