% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bulk_sample_celltype}
\alias{bulk_sample_celltype}
\title{Create pseudo-bulk for each cell type of each sample in a Seurat object}
\usage{
bulk_sample_celltype(data, ncores = 1)
}
\arguments{
\item{data}{A Seurat object containing expression.}

\item{ncores}{Number of cores for parallel computation.}
}
\value{
A Seurat object containing the pseudo-bulks for each cell type
of each sample.
}
\description{
This function takes a Seurat object as input and creates a pseudo-bulk
for each cell type of each sample in the object. This is computed by
taking the row means of the expression for each cell type of each sample
If a cell type does not exist in a sample, the expression values for that
cell type will be 0 for all genes.
}
