% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_scrnaseq}
\alias{example_scrnaseq}
\title{Example of scRNA-seq data}
\format{
\subsection{\code{example_scrnaseq}}{

A Seurat object with 3523 genes and 550 cells.
Some of the key metadata columns are:
\describe{
\item{celltype}{cell type of the cell}
\item{sample}{patient ID of the cell}
\item{condition}{whether the patient is a responder or non-responder}
}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE120575}
}
\usage{
example_scrnaseq
}
\description{
This is a subsampled version of the melanoma patients dataset as used in our manuscript.
The original dataset is available at \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE120575}.
}
\keyword{datasets}
