% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_celltype_expression.R
\name{find_var_gene}
\alias{find_var_gene}
\title{Identify the highly variable genes (HVGs) in the input data. By default,
the function calculates the HVG across the cells within each cell type,
as well across all cells). This is done for each sample separately, then
taking the union of the HVGs across all samples. The ouput is a daraframe
with two columns: marker and celltype. When celltype is set to FALSE, the
function only calculates the HVG across all cells and returns a vector of HVGs.}
\usage{
find_var_gene(
  data,
  num_top_gene = 1500,
  ncores = 1,
  celltype = TRUE,
  BPparam = NULL
)
}
\description{
Identify the highly variable genes (HVGs) in the input data. By default,
the function calculates the HVG across the cells within each cell type,
as well across all cells). This is done for each sample separately, then
taking the union of the HVGs across all samples. The ouput is a daraframe
with two columns: marker and celltype. When celltype is set to FALSE, the
function only calculates the HVG across all cells and returns a vector of HVGs.
}
