% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pathway.R
\name{format_pathway}
\alias{format_pathway}
\title{This function is used to convert the pathway score output from
gene set analysis into a matrix of samples x features. It takes
the object containing the pathway scores (in the format of pathway score x cell),
aggregates the pathway scores by cell type and converts the data into a matrix of
samples x features, where the features are the pathways and the cell types they
are associated with.}
\usage{
format_pathway(data, topMatrixGSVA, ncores)
}
\description{
This function is used to convert the pathway score output from
gene set analysis into a matrix of samples x features. It takes
the object containing the pathway scores (in the format of pathway score x cell),
aggregates the pathway scores by cell type and converts the data into a matrix of
samples x features, where the features are the pathways and the cell types they
are associated with.
}
