% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateBPParam}
\alias{generateBPParam}
\title{Enable parallel processing}
\usage{
generateBPParam(cores = 1)
}
\arguments{
\item{cores}{The number of cores to use for parallel processing.}
}
\value{
A \code{BiocParallel} object that can be used to control the
parallelization of functions.
}
\description{
This function takes the number of cores to use for parallel processing and
generates a \code{BiocParallel} object that can be used to control the
parallelization of functions.
It automatically determines whether to use the \code{SnowParam} or
\code{MulticoreParam} of the \code{BiocParallel} package based on the operating system.
}
