% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pathway.R
\name{get_geneset}
\alias{get_geneset}
\title{This function retrieves the hallmark gene set from the Molecular Signature
Database (MSigDB) for a specified species. The function returns a list of
gene sets where each list entry contains a vector of gene symbols in the gene set.}
\usage{
get_geneset(species = "Homo sapiens")
}
\description{
This function retrieves the hallmark gene set from the Molecular Signature
Database (MSigDB) for a specified species. The function returns a list of
gene sets where each list entry contains a vector of gene symbols in the gene set.
}
