% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_num_cell_per_celltype}
\alias{get_num_cell_per_celltype}
\title{Compute number of cells in each cell type in each spot
for spatial transcriptomics data}
\usage{
get_num_cell_per_celltype(data)
}
\arguments{
\item{data}{a spatial transcriptomics dataset in the form of a Seurat object.}
}
\value{
A matrix with the number of cells per cell type at each spot.
}
\description{
This function computes the number of cells in each cell type by
multiplying the cell type probability in each spot with the
relative number of cells in each spot. The relative number of cells
are estimated using library size of each spot. See get_num_cell_per_spot().
}
