% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_spatial_metrics.R
\name{helper_celltype_interaction_sp}
\alias{helper_celltype_interaction_sp}
\title{Calculates the pairwise distance between cell types for spatial proteomics.
It applies the individual_celltype_interaction_sp function to each sample,
then merges the results from each sample.}
\usage{
helper_celltype_interaction_sp(data, ncores = 1)
}
\description{
Calculates the pairwise distance between cell types for spatial proteomics.
It applies the individual_celltype_interaction_sp function to each sample,
then merges the results from each sample.
}
