% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_spatial_metrics.R
\name{helper_celltype_interaction_st}
\alias{helper_celltype_interaction_st}
\title{Calculates the cell-type interactions for spatial transcriptomic.
It applies the individual_celltype_interaction_st function to each sample,
then merges the results from each sample.}
\usage{
helper_celltype_interaction_st(data, ncores = 1)
}
\description{
Calculates the cell-type interactions for spatial transcriptomic.
It applies the individual_celltype_interaction_st function to each sample,
then merges the results from each sample.
}
