% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bulk_expression.R
\name{helper_gene_cor}
\alias{helper_gene_cor}
\title{This function calculates the correlation of gene expression in
each sample in a Seurat object. If no list of genes is provided, it will
use find_var_gene function to select the top variable genes.}
\usage{
helper_gene_cor(data, genes = NULL, num_top_gene = NULL, ncores = 1)
}
\description{
This function calculates the correlation of gene expression in
each sample in a Seurat object. If no list of genes is provided, it will
use find_var_gene function to select the top variable genes.
}
