% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bulk_expression.R
\name{helper_gene_mean}
\alias{helper_gene_mean}
\title{This function calculates the mean expression of genes in each sample in
a Seurat object. If no list of genes is provided, it will use find_var_gene
function to select the top variable genes.}
\usage{
helper_gene_mean(data, genes = NULL, num_top_gene = NULL, ncores = 1)
}
\description{
This function calculates the mean expression of genes in each sample in
a Seurat object. If no list of genes is provided, it will use find_var_gene
function to select the top variable genes.
}
