% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_celltype_expression.R
\name{helper_gene_mean_celltype}
\alias{helper_gene_mean_celltype}
\title{This function is used to calculate gene expression levels in a Seurat object
containing expression values. By default, the function first finds the variable
genes per cell type using the find_var_gene function, then calculates the gene
expression levels for these genes in their respective cell type.
The output is a returns a matrix of samples by features.}
\usage{
helper_gene_mean_celltype(
  data,
  genes = NULL,
  num_top_gene = NULL,
  ncores = 1,
  find_variable_genes = TRUE
)
}
\description{
This function is used to calculate gene expression levels in a Seurat object
containing expression values. By default, the function first finds the variable
genes per cell type using the find_var_gene function, then calculates the gene
expression levels for these genes in their respective cell type.
The output is a returns a matrix of samples by features.
}
