% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_celltype_expression.R
\name{helper_gene_mean_celltype_st}
\alias{helper_gene_mean_celltype_st}
\title{This function is used to calculate the expression of genes in each cell type
in a Seurat object containing expression values for spatial transcriptomics.
It performs a linear regression on the gene expression and cell type composition
at each spot to obtain regression coefficients and p-values. The regression
coefficients is considered as the cell type's contribution to the expression the gene.
Similar to the bulk deconvolution concept.}
\usage{
helper_gene_mean_celltype_st(data, num_top_gene = NULL, ncores = 1)
}
\description{
This function is used to calculate the expression of genes in each cell type
in a Seurat object containing expression values for spatial transcriptomics.
It performs a linear regression on the gene expression and cell type composition
at each spot to obtain regression coefficients and p-values. The regression
coefficients is considered as the cell type's contribution to the expression the gene.
Similar to the bulk deconvolution concept.
}
