% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_spatial_metrics.R
\name{helper_L_stat_st}
\alias{helper_L_stat_st}
\title{Calculates the L-statistics for spatial transcriptomic.
It applies the individual_L_stat_st function to each sample,
then merges the results from each sample.}
\usage{
helper_L_stat_st(data, ncores = 1)
}
\description{
Calculates the L-statistics for spatial transcriptomic.
It applies the individual_L_stat_st function to each sample,
then merges the results from each sample.
}
