% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_spatial_metrics.R
\name{helper_moran}
\alias{helper_moran}
\title{Calculates the Moran's I value for all samples.
It applies the individual_moran_cor function to each sample,
then merges the results from each sample.}
\usage{
helper_moran(data, num_top_gene = NULL, ncores = 1)
}
\description{
Calculates the Moran's I value for all samples.
It applies the individual_moran_cor function to each sample,
then merges the results from each sample.
}
