% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_spatial_metrics.R
\name{helper_nncorr_protein}
\alias{helper_nncorr_protein}
\title{Calculates the nearest neighbor correlation all samples
using the spatial coordinates and expression values of cells.
It applies the individual_nncorr_protein function to each sample,
then merges the results from each sample.}
\usage{
helper_nncorr_protein(data, num_top_gene = NULL, ncores = 1)
}
\description{
Calculates the nearest neighbor correlation all samples
using the spatial coordinates and expression values of cells.
It applies the individual_nncorr_protein function to each sample,
then merges the results from each sample.
}
