% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pathway.R
\name{helper_pathway_gsva}
\alias{helper_pathway_gsva}
\title{This function performs gene set enrichment analysis on a Seurat object
containing expression data to generate pathway score. The type of pathway
analysis method currently supported are ssgsea and aucell. The function
returns a matrix of samples x features with the pathway scores.}
\usage{
helper_pathway_gsva(data, method = "ssgsea", geneset, ncores = 1)
}
\description{
This function performs gene set enrichment analysis on a Seurat object
containing expression data to generate pathway score. The type of pathway
analysis method currently supported are ssgsea and aucell. The function
returns a matrix of samples x features with the pathway scores.
}
