% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pathway.R
\name{helper_pathway_mean}
\alias{helper_pathway_mean}
\title{The helper_pathway_mean function use the mean expression of genes in a
gene set as the basis of pathway score. The input is a Seurat object
containing the gene expression data, the gene set of interest. Frist
compute the mean expression of each gene in the gene set. This is
done on each cell. The mean expression values are then subtracted from
the overall mean expression of all genes in each cell to control for the
difference in library detection. The resulting gene set scores are passed
to the format_pathway function to convert them into a sample x pathway
feature matrix.}
\usage{
helper_pathway_mean(data, geneset, ncores = 1)
}
\description{
The helper_pathway_mean function use the mean expression of genes in a
gene set as the basis of pathway score. The input is a Seurat object
containing the gene expression data, the gene set of interest. Frist
compute the mean expression of each gene in the gene set. This is
done on each cell. The mean expression values are then subtracted from
the overall mean expression of all genes in each cell to control for the
difference in library detection. The resulting gene set scores are passed
to the format_pathway function to convert them into a sample x pathway
feature matrix.
}
