% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pathway.R
\name{helper_pathway_mean_st}
\alias{helper_pathway_mean_st}
\title{Thie function calculates the mean expression level of each gene set
in a list of gene sets for each cell type in each sample, for spatial
transcriptomics data. It first uses the cell type composition and
gene expression at each spot and uses linear regression to calculate the
regression coefficient of each cell type. It then sum the regression
coefficient of each cell type and returns the resulting matrix.}
\usage{
helper_pathway_mean_st(data, geneset, ncores = 1)
}
\description{
Thie function calculates the mean expression level of each gene set
in a list of gene sets for each cell type in each sample, for spatial
transcriptomics data. It first uses the cell type composition and
gene expression at each spot and uses linear regression to calculate the
regression coefficient of each cell type. It then sum the regression
coefficient of each cell type and returns the resulting matrix.
}
