% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pathway.R
\name{helper_pathway_prop}
\alias{helper_pathway_prop}
\title{Calculates proportion that a geneset is expressed in each cell type in a
sample. The function first uses the individual_geneset_proportion_celltype
function to calculate the proportion of cells expressing each gene set
in each sample. It then concatenates the gene set name to the cell type
name in the output data frame. The output data frame contains one column
for each gene set-cell type combination.}
\usage{
helper_pathway_prop(data, geneset, ncores = 1)
}
\description{
Calculates proportion that a geneset is expressed in each cell type in a
sample. The function first uses the individual_geneset_proportion_celltype
function to calculate the proportion of cells expressing each gene set
in each sample. It then concatenates the gene set name to the cell type
name in the output data frame. The output data frame contains one column
for each gene set-cell type combination.
}
