% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_proportion.R
\name{helper_proportion_ratio}
\alias{helper_proportion_ratio}
\title{This function calculates the ratio of cell type proportion between
two cell types in each sample, applicable to scRNA-seq and spatial
proteomics. The ratio is log2 transformed.}
\usage{
helper_proportion_ratio(data, ncores = 1)
}
\description{
This function calculates the ratio of cell type proportion between
two cell types in each sample, applicable to scRNA-seq and spatial
proteomics. The ratio is log2 transformed.
}
