% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_proportion.R
\name{helper_proportion_raw}
\alias{helper_proportion_raw}
\title{This function calculates the proportion of cells belonging to
each cell type in each sample, applicable to scRNA-seq and
spatial proteomics. It also takes an optional logit argument
that specifies whether the proportions should be logit-transformed.}
\usage{
helper_proportion_raw(data, logit = TRUE)
}
\description{
This function calculates the proportion of cells belonging to
each cell type in each sample, applicable to scRNA-seq and
spatial proteomics. It also takes an optional logit argument
that specifies whether the proportions should be logit-transformed.
}
