% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_proportion.R
\name{helper_proportion_raw_st}
\alias{helper_proportion_raw_st}
\title{This function calculates the proportion of cells belonging to
each cell type in each sample, applicable to spatial transcriptomics.
It takes an optional logit argument that specifies whether the proportions
should be logit-transformed.}
\usage{
helper_proportion_raw_st(data, logit = TRUE, ncores = 1)
}
\description{
This function calculates the proportion of cells belonging to
each cell type in each sample, applicable to spatial transcriptomics.
It takes an optional logit argument that specifies whether the proportions
should be logit-transformed.
}
