% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_spatial_metrics.R
\name{individual_celltype_interaction_sp}
\alias{individual_celltype_interaction_sp}
\title{This function calculates the pairwise distance between cell types
for a sample by using the coordinates and cell types of the cells.
It keeps the cell types with more than 10 cells and calculates
the pairwise distance between these cell types. The function returns
a list of the pairwise distances between the cell types.
This function is designed for spatial proteomics.}
\usage{
individual_celltype_interaction_sp(this_sample)
}
\description{
This function calculates the pairwise distance between cell types
for a sample by using the coordinates and cell types of the cells.
It keeps the cell types with more than 10 cells and calculates
the pairwise distance between these cell types. The function returns
a list of the pairwise distances between the cell types.
This function is designed for spatial proteomics.
}
