% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_spatial_metrics.R
\name{individual_celltype_interaction_st}
\alias{individual_celltype_interaction_st}
\title{This function calculates the cell-type interactions for spatial
transcriptomic. It takes probabilities of cell-type assignments
for each spot and perform matrix multiplication to get the probability of
cell-type co-occurrences for each spot. The output is a  vector containing
the co-occurrence probabilities for each cell-type pair.}
\usage{
individual_celltype_interaction_st(thisprob)
}
\description{
This function calculates the cell-type interactions for spatial
transcriptomic. It takes probabilities of cell-type assignments
for each spot and perform matrix multiplication to get the probability of
cell-type co-occurrences for each spot. The output is a  vector containing
the co-occurrence probabilities for each cell-type pair.
}
