% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_spatial_metrics.R
\name{individual_L_stat_sp}
\alias{individual_L_stat_sp}
\title{This function calculates the L-statistics for a given sample of spatial
transcriptomics data. It uses spatial coordinates, cell type information
to calculate the spatial distribution of cells in the sample.
The output vector represents the L-statistic for a pair of cell types.}
\usage{
individual_L_stat_sp(this_sample)
}
\description{
This function calculates the L-statistics for a given sample of spatial
transcriptomics data. It uses spatial coordinates, cell type information
to calculate the spatial distribution of cells in the sample.
The output vector represents the L-statistic for a pair of cell types.
}
