% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_spatial_metrics.R
\name{individual_L_stat_st}
\alias{individual_L_stat_st}
\title{This function calculates the L-statistics for a given sample for spatial
transcriptomics data using the spatial coordinates, cell type information and
the number of cells in each spatial location. The L-statistic measures the
degree of spatial clustering between two objects in a particular area. The output
vector represents the L-statistic for a pair of cell types.}
\usage{
individual_L_stat_st(thissample, this_num_cell_per_spot)
}
\description{
This function calculates the L-statistics for a given sample for spatial
transcriptomics data using the spatial coordinates, cell type information and
the number of cells in each spatial location. The L-statistic measures the
degree of spatial clustering between two objects in a particular area. The output
vector represents the L-statistic for a pair of cell types.
}
