% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_spatial_metrics.R
\name{individual_moran_cor}
\alias{individual_moran_cor}
\title{This function computes the Moran's I for each gene in a sample.
Moran's I is a measure of spatial autocorrelation, which indicates
how strongly the gene expression values in a sample cluster or disperse.}
\usage{
individual_moran_cor(thissample)
}
\description{
This function computes the Moran's I for each gene in a sample.
Moran's I is a measure of spatial autocorrelation, which indicates
how strongly the gene expression values in a sample cluster or disperse.
}
