% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_spatial_metrics.R
\name{individual_nncorr_protein}
\alias{individual_nncorr_protein}
\title{Calculates the nearest neighbor correlation for a given sample
using the spatial coordinates and expression values of cells.
It returns a vector of correlation values for each gene.
This function is used in as a helper function in helper_nncorr_protein
function to generate the nearest neighbor correlation for all samples.}
\usage{
individual_nncorr_protein(thissample)
}
\description{
Calculates the nearest neighbor correlation for a given sample
using the spatial coordinates and expression values of cells.
It returns a vector of correlation values for each gene.
This function is used in as a helper function in helper_nncorr_protein
function to generate the nearest neighbor correlation for all samples.
}
