% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{L_stats}
\alias{L_stats}
\title{Compute L statistic for a point pattern}
\usage{
L_stats(ppp_obj = NULL, from = NULL, to = NULL, L_dist = NULL)
}
\arguments{
\item{ppp_obj}{a point pattern object}

\item{from}{define the window of the point pattern to compute}

\item{to}{define the window of the point pattern to compute}

\item{L_dist}{a numeric value specifying the maximum distance
at which the L statistic will be computed.}
}
\value{
A numeric value of the L statistic.
}
\description{
This function computes the L statistic for a given point pattern.
}
