% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_data}
\alias{process_data}
\title{data pre-processing}
\usage{
process_data(data, normalise = TRUE)
}
\arguments{
\item{data}{input data, a Seurat object.}

\item{normalise}{a logical value indicating whether to normalize the data
or not. Default value is \code{TRUE}.}
}
\value{
a Seurat object
}
\description{
This function takes a Seurat object as input and does data
cleaning and pre-processing. For example, it replaces the "+" and "-"
signs in the \code{celltype} column with "plus" and "minus", respectively.
It also removes patients that have less than 10 cells across all cell types.
If the \code{normalise} argument is set to \code{TRUE}, the function will normalize
the data using the \code{Seurat::NormalizeData} function.
}
\examples{
data <- readRDS(system.file(
    "extdata", "example_spatial_transcriptomics.rds",
    package = "scFeatures"
))
data <- process_data(data, normalise = FALSE)

}
