% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rearrange_string}
\alias{rearrange_string}
\title{Rearrange string}
\usage{
rearrange_string(str)
}
\arguments{
\item{str}{A character string containing words separated by underscores.}
}
\value{
A character string with the words sorted alphabetically and
separated by underscores.
}
\description{
This function takes a string as input and rearranges words in the string
so that they are sorted alphabetically.
}
