% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_celltype_expression.R
\name{remove_mito}
\alias{remove_mito}
\title{remove mito and ribo genes and genes highly correlated with mito genes
as mitochondria and ribosomal genes may not as interesting to look at}
\usage{
remove_mito(data)
}
\arguments{
\item{data}{input data}
}
\value{
data with the mito and rib genes removed
}
\description{
remove mito and ribo genes and genes highly correlated with mito genes
as mitochondria and ribosomal genes may not as interesting to look at
}
