% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_celltype_interaction}
\alias{run_celltype_interaction}
\title{Generate cell type interaction}
\usage{
run_celltype_interaction(data, type = "spatial_p", ncores = 1)
}
\arguments{
\item{data}{A Seurat object containing \code{celltype} and \code{sample} label}

\item{type}{The type of dataset, either "scrna", "spatial_t", or "spatial_p".}

\item{ncores}{Number of cores for parallel processing.}
}
\value{
a dataframe of samples x features
}
\description{
This function calculates the pairwise distance between cell types
for a sample by using the coordinates and cell types of the cells.
The function supports spatial proteomics and spatial transcriptomics.
}
\examples{

data <- readRDS(system.file(
    "extdata", "example_spatial_proteomics.rds",
    package = "scFeatures"
))
feature_celltype_interaction <- run_celltype_interaction(
    data,
    type = "spatial_p", ncores = 1
)

}
