% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_prediction_model.R
\name{run_classification}
\alias{run_classification}
\title{run cross-validated classification}
\usage{
run_classification(X, y, model = "randomforest", ncores = 1)
}
\arguments{
\item{X}{A data matrix (i.e., observations by features)}

\item{y}{A character vector or factor containing the outcomes.}

\item{model}{Which classification model to use. Can be "randomforest", "svm"
or "dlda". Defaults to "randomforest".}

\item{ncores}{Number of cores to run cross validation on.
Defaults to 1 core.}
}
\value{
A matrix of the infile
}
\description{
This function takes a feature matrix in the form of samples by features and
performs cross-validated classification.
}
