% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_gene_cor}
\alias{run_gene_cor}
\title{generate overall aggregated gene correlation}
\usage{
run_gene_cor(
  data,
  type = "scrna",
  genes = NULL,
  num_top_gene = NULL,
  ncores = 1
)
}
\arguments{
\item{data}{input data, a Seurat object containing \code{celltype} and \code{sample} label}

\item{type}{input data type, either scrna, spatial_p, or spatial_t}

\item{genes}{default to NULL, in which case the top variable genes will be used
If provided by user, need to be in the format of a list containing the feature of interest,
eg, genes <- c(GZMA", "GZMK", "CCR7", "RPL38" )}

\item{num_top_gene}{when the genes is not provided by the user, the top variable genes will be used
The number of genes is set by this number.
default to NULL, in which case top 50 variable genes will be selected}

\item{ncores}{number of cores}
}
\value{
a matrix of samples x features
}
\description{
generate overall aggregated gene correlation
}
\examples{

data <- readRDS(system.file("extdata", "example_scrnaseq.rds", package = "scFeatures"))
feature_gene_cor <- run_gene_cor(data, type = "scrna", num_top_gene = 1500, ncores = 1)

}
