% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_gene_prop}
\alias{run_gene_prop}
\title{generate overall aggregated gene proportion expression}
\usage{
run_gene_prop(data, type = "scrna", num_top_gene = NULL, ncores = 1)
}
\arguments{
\item{data}{input data, a Seurat object containing \code{celltype} and \code{sample} label}

\item{type}{input data type, either scrna, spatial_p, or spatial_t}

\item{num_top_gene}{determines how many genes to include. By default, picks 1500 most variable genes.}

\item{ncores}{number of cores}
}
\value{
a matrix of samples x features
}
\description{
generate overall aggregated gene proportion expression
}
\examples{

data <- readRDS(system.file("extdata", "example_scrnaseq.rds", package = "scFeatures"))
feature_gene_prop <- run_gene_prop(data, type = "scrna" , num_top_gene= 1500, ncores = 1 )

}
