% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_proportion_logit}
\alias{run_proportion_logit}
\title{Generate cell type proportions, with logit transformation}
\usage{
run_proportion_logit(data, type = "scrna", ncores = 1)
}
\arguments{
\item{data}{A Seurat object containing \code{celltype} and \code{sample} label}

\item{type}{The type of dataset, either "scrna", "spatial_t", or "spatial_p".}

\item{ncores}{Number of cores for parallel processing.}
}
\value{
a matrix of samples x features
}
\description{
This function calculates the proportions of cells belonging to each cell type,
and applies a logit transformation to the proportions.
The input data must contain \code{sample} and \code{celltype} metadata column.
The function supports scRNA-seq and spatial proteomics.
The function returns a matrix with samples as rows and cell types as columns.
}
\examples{
data <- readRDS(
    system.file("extdata", "example_scrnaseq.rds", package = "scFeatures")
)
feature_proportion_logit <- run_proportion_logit(
    data,
    type = "scrna", ncores = 1
)

}
