% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_proportion_ratio}
\alias{run_proportion_ratio}
\title{Generate cell type proportion ratio}
\usage{
run_proportion_ratio(data, type = "scrna", ncores = 1)
}
\arguments{
\item{data}{A Seurat object containing \code{celltype} and \code{sample} label}

\item{type}{The type of dataset, either "scrna", "spatial_t", or "spatial_p".}

\item{ncores}{Number of cores for parallel processing.}
}
\value{
a dataframe of samples x features
}
\description{
This function calculates pairwise cell type proportion ratio for each sample.
and applies a logit transformation to the proportions.
The input data must contain \code{sample} and \code{celltype} metadata column.
The function supports scRNA-seq and spatial proteomics.
The function returns a dataframe with samples as rows and cell types as columns.
}
\examples{

data <- readRDS(
    system.file("extdata", "example_scrnaseq.rds", package = "scFeatures")
)
feature_proportion_ratio <- run_proportion_ratio(
    data,
    type = "scrna", ncores = 1
)

}
