% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_proportion_raw}
\alias{run_proportion_raw}
\title{generate cell type proportion raw}
\usage{
run_proportion_raw(data, type = "scrna", ncores = 1)
}
\arguments{
\item{data}{the input data, a Seurat object containing \code{celltype} and
\code{sample} label}

\item{type}{input data type, either scrna, spatial_p, or spatial_t}

\item{ncores}{number of cores}
}
\value{
a matrix of samples x features
}
\description{
generate cell type proportion raw
}
\examples{

data <- readRDS(
  system.file("extdata", "example_scrnaseq.rds", package = "scFeatures")
)
feature_proportion_raw <- run_proportion_raw(
  data,
  type = "scrna", ncores = 1
)

}
