% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scfeatures_result}
\alias{scfeatures_result}
\title{Example of scFeatures() output}
\format{
\subsection{\code{scfeatures_result}}{

A list with two dataframes. In each dataframe the columns are each patient and the rows are the feature values.
The first dataframe contains the feature type "proportion_raw".
The second dataframe contains the feature type "proportion_logit".
}
}
\usage{
scfeatures_result
}
\description{
This is an example output of the scFeatures() function for example_scrnaseq.
}
\keyword{datasets}
