% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_data}
\alias{check_data}
\title{check that all metadata are in the data}
\usage{
check_data(data, type = "scrna")
}
\arguments{
\item{data}{Dataset to be checked.}

\item{type}{Type of dataset (e.g., "scrna", "spatial_t", "spatial_p").}
}
\description{
check that all metadata are in the data
}
