% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_num_cell_per_spot}
\alias{get_num_cell_per_spot}
\title{estimate a relative number of cells per spot}
\usage{
get_num_cell_per_spot(data)
}
\arguments{
\item{data}{input data}
}
\value{
data with the relative number of cells/spot stored in it's metadata
}
\description{
estimate a relative number of cells per spot
}
\examples{

data <- readRDS(system.file(
 "extdata", "example_spatial_transcriptomics.rds",
                            package = "scFeatures"
))
data <- process_data(data, normalise = FALSE)
data <- get_num_cell_per_spot(data)

}
