% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_num_cell_per_spot}
\alias{get_num_cell_per_spot}
\title{Estimate a relative number of cells per spot for
spatial transcriptomics data}
\usage{
get_num_cell_per_spot(data)
}
\arguments{
\item{data}{spatial transcriptomics data in Seurat object.}
}
\value{
the object with the relative number of cells/spot stored
in the \code{number_cells} attribute.
}
\description{
This function takes a spatial transcriptomics data as input
and estimates the relative number of cells per spot in the data.
The number of cells is estimated as the library size scaled to
the range \link{1, 100}.
This value stored in the \code{number_cells} attribute.
}
\examples{

data <- readRDS(system.file(
    "extdata", "example_spatial_transcriptomics.rds",
    package = "scFeatures"
))
data <- process_data(data, normalise = FALSE)
data <- get_num_cell_per_spot(data)

}
