% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pathway.R
\name{individual_geneset_proportion_celltype}
\alias{individual_geneset_proportion_celltype}
\title{helper function to calculate proportion that a gene is expressed in each
cell type in a sample}
\usage{
individual_geneset_proportion_celltype(data, this_geneset)
}
\description{
helper function to calculate proportion that a gene is expressed in each
cell type in a sample
}
