% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_run_scfeatures.R
\name{makeSeurat}
\alias{makeSeurat}
\title{Format data into Seurat object structured for scFeatures functions}
\usage{
makeSeurat(
  data,
  sample = NULL,
  celltype = NULL,
  assay = NULL,
  spatialCoords = NULL,
  spotProbability = NULL
)
}
\arguments{
\item{data}{input data, either a SingleCellExperiment or SpatialExperiment
object.
The object needs to contain a column named "sample" and a column named
"celltype".
Alternatively, user can provide the name of the column containing sample
and celltype into the \code{sample} and \code{celltype} argument.
When passing as SingleCellExperiment or SpatialExperiment, by default we
use the assay stored in "logcount".
Alternatively, user can specify the assay to use in the \code{assay} argument.
If users want to construct features from the spatial category, by default
we need columns called "x_cord" and "y_cord".
Alternatively, please specify the relevant column in the \code{spatialCoords}
argument.
For spot-based spatial transcriptomics, we also requires a matrix
containing cell type prediction probability of each spot, in the format of
celltype x spot}

\item{sample}{a vector providing sample identifier for each cell. If not
provided, we assume the data contain a metadata column "sample" for running
scFeatures.}

\item{celltype}{a vector providing celltype identifier. If not
provided, we assume the data contain a metadata column "celltype" for
running scFeatures.}

\item{assay}{the assay identifier if using a SingleCellExperiment or
SpatialExperiment object.}

\item{spatialCoords}{the spatialCoords identifiers provided in a list of
two vectors, if users want to construct features from the spatial category.
If not provided, we assume the data contain the metadata columns "x_cord" and
"y_cord" for constructing spatial features.}

\item{spotProbability}{a matrix in the format of celltype x spot, where each
entry is the prediction probability of that cell type for each spot. This is
needed by spatial transcriptomics data.}
}
\value{
A \code{Seurat} dataset containing required metadata for running scFeatures.
}
\description{
This function is used to convert a SingleCellExperiment, SpatialExperiment or a
Seurat object into Seurat object containing all required fields and structured
for scFeatures functions.
}
\examples{

 data <- readRDS(system.file("extdata", "example_scrnaseq.rds", package = "scFeatures"))
 coordinate <- list(x = rep(1, ncol(data)), y = rep(1, ncol(data)))
 data <- makeSeurat(data, spatialCoords = coordinate)

}
