% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_association_study_report}
\alias{run_association_study_report}
\title{automatically generate the association study report in an html format}
\usage{
run_association_study_report(scfeatures_result, output_folder)
}
\arguments{
\item{scfeatures_result}{A named list storing the scFeatures feature output}

\item{output_folder}{A directory for saving the html report file}
}
\value{
the html file will be saved in the directory defined in the output
folder
}
\description{
automatically generate the association study report in an html format
}
