% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_CCI}
\alias{run_CCI}
\title{generate cell cell communication score}
\usage{
run_CCI(data, species = "Homo sapiens", type = "scrna", ncores = 8)
}
\arguments{
\item{data}{input data}

\item{type}{input data type, either scrna, spatial_p, or spatial_t}

\item{ncores}{number of cores}
}
\value{
a matrix of samples x features
}
\description{
generate cell cell communication score
}
