% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_CCI}
\alias{run_CCI}
\title{Generate cell cell communication score}
\usage{
run_CCI(data, type = "scrna", ncores = 1)
}
\arguments{
\item{data}{input data, a Seurat object containing \code{celltype} and \code{sample} label}

\item{type}{input data type, either scrna, spatial_p, or spatial_t}

\item{ncores}{number of cores}
}
\value{
a matrix of samples x features
}
\description{
This function calculates the ligand receptor interaction score using SingleCellSignalR.
The output features are in the form of celltype a -> celltype b -- ligand 1 -> receptor 2 ,
which indicates the interaction between ligand 1 in celltype a and receptor 2 in celltype b.

It supports scRNA-seq.
}
\examples{

data <- readRDS(
    system.file("extdata", "example_scrnaseq.rds", package = "scFeatures")
)[1:50, 1:20]
feature_CCI <- run_CCI(data, type = "scrna" ,  ncores = 1 )

}
