% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_celltype_interaction}
\alias{run_celltype_interaction}
\title{generate cell type interaction}
\usage{
run_celltype_interaction(data, type = "spatial_p", ncores = 8)
}
\arguments{
\item{data}{input data}

\item{type}{input data type, either scrna, spatial_p, or spatial_t}

\item{ncores}{number of cores}
}
\value{
a matrix of samples x features
}
\description{
generate cell type interaction
}
