% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_prediction_model.R
\name{run_classification}
\alias{run_classification}
\title{run cross-validated classification}
\usage{
run_classification(X, y, model = "randomforest", ncores = 8)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function takes a feature matrix in the form of samples by features and performs cross-validated classification.
}
