% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_gene_cor}
\alias{run_gene_cor}
\title{generate bulk gene correlation}
\usage{
run_gene_cor(data, type = "scrna", num_top_gene = NULL, ncores = 8)
}
\arguments{
\item{data}{input data}

\item{type}{input data type, either scrna, spatial_p, or spatial_t}

\item{ncores}{number of cores}
}
\value{
a matrix of samples x features
}
\description{
generate bulk gene correlation
}
