% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_gene_prop}
\alias{run_gene_prop}
\title{Generate overall aggregated gene proportion expression}
\usage{
run_gene_prop(
  data,
  type = "scrna",
  genes = NULL,
  num_top_gene = NULL,
  ncores = 1
)
}
\arguments{
\item{data}{A Seurat object containing \code{celltype} and \code{sample} label}

\item{type}{The type of dataset, either "scrna", "spatial_t", or "spatial_p".}

\item{genes}{Default to NULL, in which case the top variable genes will be
used. If provided by user, need to be in the format of a list containing the
genes of interest, eg, genes <- c(GZMA", "GZMK", "CCR7", "RPL38" )}

\item{num_top_gene}{Number of top variable genes to use when genes is not provided.
Defaults to 1500.}

\item{ncores}{Number of cores for parallel processing.}
}
\value{
a dataframe of samples x features
}
\description{
This function computes the proportion of gene expression across samples. The user
can specify the genes of interest, or let the function use the top variable
genes by default. The function supports scRNA-seq, spatial proteomics,
and spatial transcriptomics.
}
\examples{

data <- readRDS(system.file("extdata", "example_scrnaseq.rds", package = "scFeatures"))
feature_gene_prop <- run_gene_prop(data, type = "scrna", num_top_gene = 1500, ncores = 1)

}
