% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_gene_prop_celltype}
\alias{run_gene_prop_celltype}
\title{generate cell type specific gene proportion expression}
\usage{
run_gene_prop_celltype(
  data,
  type = "scrna",
  genes = NULL,
  num_top_gene = NULL,
  ncores = 1
)
}
\arguments{
\item{data}{input data, a Seurat object containing \code{celltype} and \code{sample} label}

\item{type}{input data type, either scrna, spatial_p, or spatial_t}

\item{genes}{default to NULL, in which case the top variable genes will be used
If provided by user, need to be in the format of a dataframe with two columns 'marker' and 'celltype'.
The marker column contains the genes of interest, eg: S100A11 , CCL4 ,
the celltype column contains the celltype that the gene expression is to be computed from, eg: CD8, B cells}

\item{num_top_gene}{when the genes is not provided by the user, the top variable genes will be used
The number of genes is set by this number.
default to NULL, in which case top 100 genes from each cell type will be selected}

\item{ncores}{number of cores}
}
\value{
a matrix of samples x features
}
\description{
generate cell type specific gene proportion expression
}
\examples{

data <- readRDS(
  system.file("extdata", "example_scrnaseq.rds", package = "scFeatures")
)
# optional step, if mito and ribo genes are not of interest
data_remove_mito <- remove_mito(data)
feature_gene_prop_celltype <- run_gene_prop_celltype(
  data_remove_mito,
  type = "scrna", num_top_gene = 100, ncores = 1
)

}
