% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_gene_prop_celltype}
\alias{run_gene_prop_celltype}
\title{generate cell type specific gene proportion expression}
\usage{
run_gene_prop_celltype(data, type = "scrna", num_top_gene = NULL, ncores = 8)
}
\arguments{
\item{data}{input data}

\item{type}{input data type, either scrna, spatial_p, or spatial_t}

\item{ncores}{number of cores}
}
\value{
a matrix of samples x features
}
\description{
generate cell type specific gene proportion expression
}
