% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_L_function}
\alias{run_L_function}
\title{Generate L stats}
\usage{
run_L_function(data, type = "spatial_p", ncores = 1)
}
\arguments{
\item{data}{A Seurat object containing \code{celltype} and \code{sample} label}

\item{type}{The type of dataset, either "scrna", "spatial_t", or "spatial_p".}

\item{ncores}{Number of cores for parallel processing.}
}
\value{
a dataframe of samples x features
}
\description{
This function calculates L-statistics to measure spatial autocorrelation.
The function supports  spatial proteomics and spatial transcriptomics.
}
\examples{

data <- readRDS(
system.file("extdata", "example_scrnaseq.rds", package = "scFeatures")
)
x <- sample(1:100, ncol(data) , replace = TRUE)
y <- sample(1:100, ncol(data) , replace = TRUE)
data <- makeSeurat(data, spatialCoords = list(x,y))

feature_L_function <- run_L_function(data, type = "spatial_p", ncores = 1)

}
