% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_Morans_I}
\alias{run_Morans_I}
\title{generate Moran's I}
\usage{
run_Morans_I(data, type = "spatial_p", ncores = 8)
}
\arguments{
\item{data}{input data}

\item{type}{input data type, either scrna, spatial_p, or spatial_t}

\item{ncores}{number of cores}
}
\value{
a matrix of samples x features
}
\description{
generate Moran's I
}
