% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_Morans_I}
\alias{run_Morans_I}
\title{Generate Moran's I}
\usage{
run_Morans_I(data, type = "spatial_p", ncores = 1)
}
\arguments{
\item{data}{A Seurat object containing \code{celltype} and \code{sample} label}

\item{type}{The type of dataset, either "scrna", "spatial_t", or "spatial_p".}

\item{ncores}{Number of cores for parallel processing.}
}
\value{
a dataframe of samples x features
}
\description{
This function calculates Moran's I to measure spatial autocorrelation,
which an indicattion of how strongly the feature(ie, genes/proteins)
expression values in a sample cluster or disperse. The function supports
spatial proteomics and spatial transcriptomics.
}
\examples{

data("example_scrnaseq" , package = "scFeatures")
data <- example_scrnaseq[1:50, 1:20]

x <- sample(1:100, ncol(data) , replace = TRUE)
y <- sample(1:100, ncol(data) , replace = TRUE)
data <- makeSeurat(data, spatialCoords = list(x,y))
data$sample <- sample( c("patient1", "patient2", "patient3"), ncol(data) , replace= TRUE )

feature_celltype_interaction <- run_celltype_interaction(
    data,
    type = "spatial_p", ncores = 1
)

feature_Morans_I <- run_Morans_I(data, type = "spatial_p", ncores = 1)

}
