% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scfeatures.R
\name{run_pathway_gsva}
\alias{run_pathway_gsva}
\title{generate pathway score using gene set enrichement analysis}
\usage{
run_pathway_gsva(
  data,
  geneset = NULL,
  species = "Homo sapiens",
  type = "scrna",
  subsample = T,
  ncores = 8
)
}
\arguments{
\item{data}{input data}

\item{type}{input data type, either scrna, spatial_p, or spatial_t}

\item{ncores}{number of cores}
}
\value{
a matrix of samples x features
}
\description{
generate pathway score using gene set enrichement analysis
}
