% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_run_scfeatures.R
\name{scFeatures}
\alias{scFeatures}
\title{wrapper function to run all feature types in scFeatures}
\usage{
scFeatures(
  data,
  feature_types = NULL,
  type = "scrna",
  ncores = 1,
  species = "Homo sapiens",
  celltype_genes = NULL,
  aggregated_genes = NULL,
  geneset = NULL,
  sample = "sample",
  celltype = "celltype",
  assay = "logcounts",
  spatialCoords = NULL
)
}
\arguments{
\item{data}{input data, a Seurat object containing "sample" and "celltype" column.
"x_cord" and "y_cord" is also required for constructing the features in the spatial metrics category.}

\item{feature_types}{vector containing the name of the feature types to generate,
options are "proportion_raw", "proportion_logit" , "proportion_ratio",
"gene_mean_celltype", "gene_prop_celltype", "gene_cor_celltype",
"pathway_gsva" , "pathway_mean", "pathway_prop" , "CCI",
"gene_mean_aggregated", "gene_prop_aggregated", 'gene_cor_aggregated',
"L_stats" , "celltype_interaction" , "morans_I", "nn_correlation".
If no value is provided, all the above feature types will be generated.}

\item{type}{input data type, either "scrna" (stands for single-cell RNA-sequencing data),
"spatial_p" (stands for spatial proteomics data), or "spatial_t" (stands for single cell spatial data )}

\item{ncores}{number of cores , default to 1}

\item{species}{either "Homo sapiens" or "Mus musculus". Defaults to "Homo sapiens" if no value provided}

\item{celltype_genes}{the genes of interest for celltype specific gene expression feature category
If no value is provided, the top variable genes will be used}

\item{aggregated_genes}{the genes of interest for overall aggregated gene expression feature category
If no value is provided, the top variable genes will be used}

\item{geneset}{the geneset of interest for celltype specific pathway feature category
If no value is provided, the 50 hallmark pathways will be used}

\item{sample}{the sample identifier if using a SingleCellExperiment}

\item{celltype}{the celltype identifier if using a SingleCellExperiment}

\item{assay}{the assay identifier if using a SingleCellExperiment}

\item{spatialCoords}{the spatialCoords identifiers if using a SingleCellExperiment}
}
\value{
a list of dataframes containing the generated feature matrix in the form of sample x features
}
\description{
wrapper function to run all feature types in scFeatures
}
\examples{

data <- readRDS(system.file("extdata", "example_scrnaseq.rds", package = "scFeatures"))
scfeatures_result <- scFeatures(data, type = "scrna" )

}
